+++ trunk/src/jpcsp/graphics/RE/REShader.java
@@ -628,7 +628,7 @@ public class REShader extends BaseRenderingEngineFunction {
 	}
 
 	@Override
-	public void setDepthRange(float zpos, float zscale, float near, float far) {
+	public void setDepthRange(float zpos, float zscale, int near, int far) {
 		shaderContext.setZPos(zpos);
 		shaderContext.setZScale(zscale);
 		super.setDepthRange(zpos, zscale, near, far);
